# feat(bench): improve speed bench output