--- a/lib/speed.js
+++ b/lib/speed.js
@@ -2,7 +2,8 @@
 
 const util = require('util');
 const { iter } = require('@metarhia/common');
-const { roundTo } = require('../lib/utils');
+const { roundTo, sum } = require('../lib/utils');
+const { computeStats } = require('./stats');
 
 const NS_PER_SEC = 1e9;
 
@@ -96,6 +97,22 @@ const measure = (
   return res;
 };
 
+const aggregateResults = results =>
+  iter(results)
+    .chainApply(it => it.groupBy(t => t.name))
+    .map(([name, samples]) => {
+      const stats = computeStats(
+        samples.map(s => s.time / NS_PER_SEC / s.count)
+      );
+      const time = sum(iter(samples).map(s => s.time)) / samples.length;
+      const ops = 1 / stats.mean;
+      return { name, samples, stats, time, ops };
+    })
+    .toArray()
+    // TODO(lundibundi): use better test comparison like
+    //  http://www.statisticslectures.com/topics/mannwhitneyu
+    .sort((t1, t2) => t2.ops - t1.ops);
+
 // Microbenchmark each passed function and compare results.
 //   caption <string> name of the benchmark
 //   count <number> amount of times ro run each function
@@ -103,7 +120,7 @@ const measure = (
 const speed = (caption, count, cases) => {
   const times = measure(cases, {
     defaultCount: count,
-    runs: 1,
+    runs: 5,
     preflight: 3,
     listener: {
       preflight: name => console.log(`start preflight ${name}`),
@@ -111,15 +128,19 @@ const speed = (caption, count, cases) => {
       done: name => console.log(`done ${name}`),
     },
   });
-  const ladder = times.sort((t1, t2) => t1.time - t2.time);
+  const ladder = aggregateResults(times);
   const best = ladder[0].time;
-  console.log(caption);
+  console.log('\n' + caption);
   ladder.forEach(test => {
-    const name = test.name.padEnd(25, '.');
-    const time = test.time.toString().padStart(15, '.');
-    const percent = roundTo((test.time * 100) / best - 100, 2).toString();
-    const percentstr = (percent > 0 ? `+${percent}%` : 'min').padStart(10, '.');
-    console.log(name + time + percentstr);
+    const name = test.name.padEnd(15, ' ');
+    const time = Math.round(test.time) + 'ns ';
+    const percent = roundTo((test.time * 100) / best - 100, 2);
+    const percentstr = (percent > 0 ? `+${percent}%` : 'min')
+      .padStart(5, ' ')
+      .padEnd(15, ' ');
+    const ops = roundTo(test.ops, 2) + ' ops/s ';
+    const rme = '±' + roundTo(test.stats.relativeMarginOfError, 2) + '%';
+    console.log(name + time + percentstr + ops + rme);
   });
   console.log();
 };