--- a/lib/speed.js
+++ b/lib/speed.js
@@ -47,9 +47,14 @@ const run = (count, fn, args = []) => {
 //         name <string> case name
 //         count <number> number of times it will be run
 //         args <Array> function arguments
-//       done <Function> called when run in done, optional
+//       cycle <Function> called when run is done, optional
 //         name <string> case name
-//         results <Object> case run results
+//         result <Object> case results
+//       done <Function> called when all runs for given configurations are done,
+//           optional
+//         name <string> case name
+//         args <Array> current configuration
+//         results <Array> results of all runs with this configuration
 //       finish <Function> called when measuring is finished, optional
 //         results <Array> all case results
 // Returns: <Array> results of all cases as objects of structure
@@ -84,13 +89,16 @@ const measure = (
       }
     }
     for (const args of argCases) {
+      const argResults = [];
       for (let i = 0; i < runs; i++) {
         if (listener.run) listener.run(testName, n, args);
         const r = run(n, fn, args);
         const result = { ...r, name: testName };
-        if (listener.done) listener.done(testName, result);
-        res.push(result);
+        if (listener.cycle) listener.cycle(testName, result);
+        argResults.push(result);
       }
+      if (listener.done) listener.done(testName, args, argResults);
+      res.push(...argResults);
     }
   });
   if (listener.finish) listener.finish(res);
@@ -190,4 +198,5 @@ module.exports = {
   resultToCsv,
   convertToCsv,
   makeTotalResults,
+  aggregateResults,
 };