--- a/lib/utils.js
+++ b/lib/utils.js
@@ -50,6 +50,8 @@ const roundTo = (num, decimals) => {
   return Math.round(num * c) / c;
 };
 
+const sum = reducable => reducable.reduce((acc, v) => acc + v, 0);
+
 module.exports = {
   inspect,
   format,
@@ -60,4 +62,5 @@ module.exports = {
   compareFirstKey,
   escapeRegex,
   roundTo,
+  sum,
 };