… lib/stats.js