New `speed` output:
```cmd
$ metatests speed benchmarks/cases/setVsArray.js --target size5.match
...
Speed test
searchSet      792317500ns   min          12621202.98 ops/s ±1.64%
searchArray    824139121ns +4.02%         12133873.69 ops/s ±1.19%
```

Old `speed` output
```cmd
$ metatests speed benchmarks/cases/setVsArray.js --target size5.match
...
Speed test
searchSet......................726039154.......min
searchArray...................1351735530...+86.18%
```
---------------------------------
New `measure` output (the old output is still available with `--verbose`)
```cmd
$ metatests measure benchmarks/example/create.js
defineObject   32972630.18 ops/s ±3.85%
mixinObject    34238572.84 ops/s ±1.72%
```
Old `measure` output
```cmd
$ metatests measure benchmarks/example/create.js
defineObject    n=1000000:              27190838ns, 36777093.81ops/s
defineObject    n=1000000:              30433772ns, 32858233.94ops/s
defineObject    n=1000000:              30466257ns, 32823198.47ops/s
defineObject    n=1000000:              30604932ns, 32674472.21ops/s
defineObject    n=1000000:              31591362ns, 31654222.44ops/s
defineObject    n=1000000:              30758265ns, 32511586.72ops/s
defineObject    n=1000000:              30378491ns, 32918027.43ops/s
defineObject    n=1000000:              29996100ns, 33337667.23ops/s
defineObject    n=1000000:              30555639ns, 32727183.35ops/s
defineObject    n=1000000:              31740639ns, 31505351.86ops/s
defineObject    n=1000000:              31576109ns, 31669513.18ops/s
defineObject    n=1000000:              30549053ns, 32734238.93ops/s
defineObject    n=1000000:              29693945ns, 33676899.45ops/s
defineObject    n=1000000:              30464988ns, 32824565.69ops/s
defineObject    n=1000000:              30087619ns, 33236262.4ops/s
defineObject    n=1000000:              31392956ns, 31854279.67ops/s
defineObject    n=1000000:              29608957ns, 33773563.86ops/s
defineObject    n=1000000:              40608643ns, 24625299.59ops/s
defineObject    n=1000000:              40201016ns, 24874993.21ops/s
defineObject    n=1000000:              33822013ns, 29566542.95ops/s
mixinObject     n=1000000:              29733543ns, 33632049.84ops/s
mixinObject     n=1000000:              33056377ns, 30251349.08ops/s
mixinObject     n=1000000:              35438258ns, 28218091.31ops/s
mixinObject     n=1000000:              32281742ns, 30977262.63ops/s
mixinObject     n=1000000:              33958784ns, 29447461.96ops/s
mixinObject     n=1000000:              32509604ns, 30760140.91ops/s
mixinObject     n=1000000:              31235763ns, 32014585.33ops/s
mixinObject     n=1000000:              31415716ns, 31831201.94ops/s
mixinObject     n=1000000:              31566123ns, 31679531.88ops/s
mixinObject     n=1000000:              31124214ns, 32129325.42ops/s
mixinObject     n=1000000:              29919812ns, 33422669.9ops/s
mixinObject     n=1000000:              30891851ns, 32370996.48ops/s
mixinObject     n=1000000:              35654862ns, 28046665.84ops/s
mixinObject     n=1000000:              32572663ns, 30700590.86ops/s
mixinObject     n=1000000:              34478141ns, 29003883.94ops/s
mixinObject     n=1000000:              33451602ns, 29893934.53ops/s
mixinObject     n=1000000:              32389873ns, 30873847.51ops/s
mixinObject     n=1000000:              32102164ns, 31150547.98ops/s
mixinObject     n=1000000:              31317254ns, 31931279.8ops/s
mixinObject     n=1000000:              31230215ns, 32020272.68ops/s
```